netstack/gro: cast Checksum arg to uint64 for amneziawg-go API
amneziawg-go's exported tun.Checksum takes uint64 for the accumulator; the
Tailscale fork used uint16. One-line cast to keep the call site compiling
against the new drop-in dependency. Also gitignore .omc/ state directory.

+++ wgengine/netstack/gro/gro.go
@@ -87,7 +87,7 @@ func RXChecksumOffload(p *packet.Parsed) *stack.PacketBuffer {
 			p.Src.Addr().AsSlice(),
 			p.Dst.Addr().AsSlice(),
 			uint16(lenForPseudo))
-		csum = tun.Checksum(buf[csumStart:], csum)
+		csum = tun.Checksum(buf[csumStart:], uint64(csum))
 		if ^csum != 0 {
 			return nil
 		}
